+++ packages/lib/test/ref/rootRef.test.ts
@@ -588,6 +588,10 @@ test("undo manager can undo removal of a referenced object in a single step", ()
     Array [
       Object {
         "actionName": "setSelectedCountry",
+        "attachedState": Object {
+          "afterEvent": undefined,
+          "beforeEvent": undefined,
+        },
         "inversePatches": Array [
           Object {
             "op": "replace",
@@ -625,6 +629,10 @@ test("undo manager can undo removal of a referenced object in a single step", ()
     Array [
       Object {
         "actionName": "removeCountry",
+        "attachedState": Object {
+          "afterEvent": undefined,
+          "beforeEvent": undefined,
+        },
         "inversePatches": Array [
           Object {
             "op": "add",